Make the data upload max size fields configurable (#13)
Currently we're using the django admin to add additional fields, because that portion of the API hasn't been built yet. The issue there is that, Django has a default maximum number of fields that it allows in a POST request, and on some flows we've exceeded that, so we can no longer add any additional fields.

This PR makes that value configurable, so that we can set it higher in the places where we need it to be higher.